Make HALCaller config more expressive
